+++ examples/apikeys/apikeys.go
@@ -43,6 +43,9 @@ func RetrieveallAPIKeysbelongingtotheauthenticateduser() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/api_keys", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["limit"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)