+++ examples/browsers/browsers.go
@@ -15,6 +15,14 @@ func Retrieveemailstatisticsbybrowser() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/browsers/stats", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["end_date"] = "2016-04-01"
+  queryParams["aggregated_by"] = "day"
+  queryParams["browsers"] = "test_string"
+  queryParams["limit"] = "test_string"
+  queryParams["offset"] = "test_string"
+  queryParams["start_date"] = "2016-01-01"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)