+++ examples/campaigns/campaigns.go
@@ -54,6 +54,10 @@ func RetrieveallCampaigns() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/campaigns", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)