+++ examples/categories/categories.go
@@ -15,6 +15,11 @@ func Retrieveallcategories() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/categories", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["category"] = "test_string"
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -34,6 +39,14 @@ func RetrieveEmailStatisticsforCategories() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/categories/stats", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["end_date"] = "2016-04-01"
+  queryParams["aggregated_by"] = "day"
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  queryParams["start_date"] = "2016-01-01"
+  queryParams["categories"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -53,6 +66,15 @@ func Retrievesumsofemailstatsforeachcategory() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/categories/stats/sums", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["end_date"] = "2016-04-01"
+  queryParams["aggregated_by"] = "day"
+  queryParams["limit"] = "1"
+  queryParams["sort_by_metric"] = "test_string"
+  queryParams["offset"] = "1"
+  queryParams["start_date"] = "2016-01-01"
+  queryParams["sort_by_direction"] = "asc"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)