+++ examples/contactdb/contactdb.go
@@ -164,6 +164,9 @@ func UpdateaList() {
   request.Body = []byte(` {
   "name": "newlistname"
 }`)
+  queryParams := make(map[string]string)
+  queryParams["list_id"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -183,6 +186,9 @@ func Retrieveasinglelist() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/lists/{list_id}", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["list_id"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -202,6 +208,9 @@ func DeleteaList() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/lists/{list_id}", host)
   request.Method = "DELETE"
+  queryParams := make(map[string]string)
+  queryParams["delete_contacts"] = "true"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -244,6 +253,11 @@ func RetrieveallrecipientsonaList() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/lists/{list_id}/recipients", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["page"] = "1"
+  queryParams["page_size"] = "1"
+  queryParams["list_id"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -282,6 +296,10 @@ func DeleteaSingleRecipientfromaSingleList() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/lists/{list_id}/recipients/{recipient_id}", host)
   request.Method = "DELETE"
+  queryParams := make(map[string]string)
+  queryParams["recipient_id"] = "1"
+  queryParams["list_id"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -360,6 +378,10 @@ func Retrieverecipients() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/recipients", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["page"] = "1"
+  queryParams["page_size"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -440,6 +462,9 @@ func Retrieverecipientsmatchingsearchcriteria() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/recipients/search", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["{field_name}"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -609,6 +634,9 @@ func Updateasegment() {
   "list_id": 5, 
   "name": "The Millers"
 }`)
+  queryParams := make(map[string]string)
+  queryParams["segment_id"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -628,6 +656,9 @@ func Retrieveasegment() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/segments/{segment_id}", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["segment_id"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -647,6 +678,9 @@ func Deleteasegment() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/segments/{segment_id}", host)
   request.Method = "DELETE"
+  queryParams := make(map[string]string)
+  queryParams["delete_contacts"] = "true"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -666,6 +700,10 @@ func Retrieverecipientsonasegment() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/contactdb/segments/{segment_id}/recipients", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["page"] = "1"
+  queryParams["page_size"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)