+++ examples/devices/devices.go
@@ -15,6 +15,13 @@ func Retrieveemailstatisticsbydevicetype() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/devices/stats", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["aggregated_by"] = "day"
+  queryParams["limit"] = "1"
+  queryParams["start_date"] = "2016-01-01"
+  queryParams["end_date"] = "2016-04-01"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)