+++ examples/subusers/subusers.go
@@ -43,6 +43,11 @@ func ListallSubusers() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/subusers", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["username"] = "test_string"
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -62,6 +67,9 @@ func RetrieveSubuserReputations() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/subusers/reputations", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["usernames"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -81,6 +89,14 @@ func Retrieveemailstatisticsforyoursubusers() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/subusers/stats", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["end_date"] = "2016-04-01"
+  queryParams["aggregated_by"] = "day"
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  queryParams["start_date"] = "2016-01-01"
+  queryParams["subusers"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -100,6 +116,14 @@ func Retrievemonthlystatsforallsubusers() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/subusers/stats/monthly", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["subuser"] = "test_string"
+  queryParams["limit"] = "1"
+  queryParams["sort_by_metric"] = "test_string"
+  queryParams["offset"] = "1"
+  queryParams["date"] = "test_string"
+  queryParams["sort_by_direction"] = "asc"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -119,6 +143,15 @@ func Retrievethetotalsforeachemailstatisticmetricforallsubusers() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/subusers/stats/sums", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["end_date"] = "2016-04-01"
+  queryParams["aggregated_by"] = "day"
+  queryParams["limit"] = "1"
+  queryParams["sort_by_metric"] = "test_string"
+  queryParams["offset"] = "1"
+  queryParams["start_date"] = "2016-01-01"
+  queryParams["sort_by_direction"] = "asc"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -285,6 +318,13 @@ func Retrievethemonthlyemailstatisticsforasinglesubuser() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/subusers/{subuser_name}/stats/monthly", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["date"] = "test_string"
+  queryParams["sort_by_direction"] = "asc"
+  queryParams["limit"] = "1"
+  queryParams["sort_by_metric"] = "test_string"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)