+++ examples/whitelabel/whitelabel.go
@@ -46,6 +46,13 @@ func Listalldomainwhitelabels() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/whitelabel/domains", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["username"] = "test_string"
+  queryParams["domain"] = "test_string"
+  queryParams["exclude_subusers"] = "true"
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -289,6 +296,11 @@ func RetrieveallIPwhitelabels() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/whitelabel/ips", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["ip"] = "test_string"
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -370,6 +382,10 @@ func CreateaLinkWhitelabel() {
   "domain": "example.com", 
   "subdomain": "mail"
 }`)
+  queryParams := make(map[string]string)
+  queryParams["limit"] = "1"
+  queryParams["offset"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -389,6 +405,9 @@ func Retrievealllinkwhitelabels() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/whitelabel/links", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["limit"] = "1"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -408,6 +427,9 @@ func RetrieveaDefaultLinkWhitelabel() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/whitelabel/links/default", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["domain"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -427,6 +449,9 @@ func RetrieveAssociatedLinkWhitelabel() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/whitelabel/links/subuser", host)
   request.Method = "GET"
+  queryParams := make(map[string]string)
+  queryParams["username"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)
@@ -446,6 +471,9 @@ func DisassociateaLinkWhitelabel() {
   host := "https://api.sendgrid.com"
   request := sendgrid.GetRequest(apiKey, "/v3/whitelabel/links/subuser", host)
   request.Method = "DELETE"
+  queryParams := make(map[string]string)
+  queryParams["username"] = "test_string"
+  request.QueryParams = queryParams
   response, err := sendgrid.API(request)
   if err != nil {
     fmt.Println(err)